+++ src/ray/core_worker/core_worker.cc
@@ -87,8 +87,7 @@ CoreWorker::CoreWorker(const WorkerType worker_type, const Language language,
       memory_store_(std::make_shared<CoreWorkerMemoryStore>()),
       task_execution_service_work_(task_execution_service_),
       task_execution_callback_(task_execution_callback),
-      task_grpc_service_(task_execution_service_, *this),
-      direct_task_grpc_service_(task_execution_service_, *this) {
+      task_grpc_service_(task_execution_service_, *this) {
   // Initialize logging if log_dir is passed. Otherwise, it must be initialized
   // and cleaned up by the caller.
   if (log_dir_ != "") {
@@ -115,12 +114,11 @@ CoreWorker::CoreWorker(const WorkerType worker_type, const Language language,
                                   std::placeholders::_2, std::placeholders::_3);
     raylet_task_receiver_ = std::unique_ptr<CoreWorkerRayletTaskReceiver>(
         new CoreWorkerRayletTaskReceiver(raylet_client_, execute_task));
-    worker_server_.RegisterService(task_grpc_service_);
     direct_actor_task_receiver_ = std::unique_ptr<CoreWorkerDirectActorTaskReceiver>(
         new CoreWorkerDirectActorTaskReceiver(worker_context_, task_execution_service_,
                                               worker_server_, execute_task,
                                               exit_handler));
-    worker_server_.RegisterService(direct_task_grpc_service_);
+    worker_server_.RegisterService(task_grpc_service_);
   }
 
   // Start RPC server after all the task receivers are properly initialized.
