+++ src/ray/core_worker/transport/direct_actor_transport.h
@@ -14,8 +14,8 @@
 #include "ray/core_worker/context.h"
 #include "ray/core_worker/store_provider/memory_store_provider.h"
 #include "ray/gcs/redis_gcs_client.h"
-#include "ray/rpc/worker/direct_actor_client.h"
-#include "ray/rpc/worker/direct_actor_server.h"
+#include "ray/rpc/grpc_server.h"
+#include "ray/rpc/worker/worker_client.h"
 
 namespace ray {
 
@@ -67,7 +67,7 @@ class CoreWorkerDirectActorTaskSubmitter {
   /// \param[in] task_id The ID of a task.
   /// \param[in] num_returns Number of return objects.
   /// \return Void.
-  void PushTask(rpc::DirectActorClient &client,
+  void PushTask(rpc::WorkerTaskClient &client,
                 std::unique_ptr<rpc::PushTaskRequest> request, const ActorID &actor_id,
                 const TaskID &task_id, int num_returns);
 
@@ -114,7 +114,7 @@ class CoreWorkerDirectActorTaskSubmitter {
   ///
   /// TODO(zhijunfu): this will be moved into `actor_states_` later when we can
   /// subscribe updates for a specific actor.
-  std::unordered_map<ActorID, std::shared_ptr<rpc::DirectActorClient>> rpc_clients_;
+  std::unordered_map<ActorID, std::shared_ptr<rpc::WorkerTaskClient>> rpc_clients_;
 
   /// Map from actor id to the actor's pending requests.
   std::unordered_map<ActorID, std::list<std::unique_ptr<rpc::PushTaskRequest>>>
@@ -327,7 +327,7 @@ class SchedulingQueue {
   friend class SchedulingQueueTest;
 };
 
-class CoreWorkerDirectActorTaskReceiver : public rpc::DirectActorHandler {
+class CoreWorkerDirectActorTaskReceiver {
  public:
   using TaskHandler = std::function<Status(
       const TaskSpecification &task_spec, const ResourceMappingType &resource_ids,
@@ -348,7 +348,7 @@ class CoreWorkerDirectActorTaskReceiver : public rpc::DirectActorHandler {
   /// \param[out] reply The reply message.
   /// \param[in] send_reply_callback The callback to be called when the request is done.
   void HandlePushTask(const rpc::PushTaskRequest &request, rpc::PushTaskReply *reply,
-                      rpc::SendReplyCallback send_reply_callback) override;
+                      rpc::SendReplyCallback send_reply_callback);
 
   /// Handle a `DirectActorCallArgWaitComplete` request.
   ///
@@ -358,7 +358,7 @@ class CoreWorkerDirectActorTaskReceiver : public rpc::DirectActorHandler {
   void HandleDirectActorCallArgWaitComplete(
       const rpc::DirectActorCallArgWaitCompleteRequest &request,
       rpc::DirectActorCallArgWaitCompleteReply *reply,
-      rpc::SendReplyCallback send_reply_callback) override;
+      rpc::SendReplyCallback send_reply_callback);
 
   /// Set the max concurrency at runtime. It cannot be changed once set.
   void SetMaxActorConcurrency(int max_concurrency);
