+++ src/ray/protobuf/worker.proto
@@ -44,12 +44,6 @@ message AssignTaskRequest {
 message AssignTaskReply {
 }
 
-// Service for worker.
-service WorkerTaskService {
-  // Push a task to a worker.
-  rpc AssignTask(AssignTaskRequest) returns (AssignTaskReply);
-}
-
 message ReturnObject {
   // Object ID.
   bytes object_id = 1;
@@ -88,9 +82,11 @@ message DirectActorCallArgWaitCompleteRequest {
 message DirectActorCallArgWaitCompleteReply {
 }
 
-// Service for direct actor.
-service DirectActorService {
+// Service for worker.
+service WorkerTaskService {
   // Push a task to a worker.
+  rpc AssignTask(AssignTaskRequest) returns (AssignTaskReply);
+  // Push a task to a direct-call actor.
   rpc PushTask(PushTaskRequest) returns (PushTaskReply);
   // Notify wait for direct actor call args has completed
   rpc DirectActorCallArgWaitComplete(DirectActorCallArgWaitCompleteRequest)
