+++ src/ray/raylet/worker.h
@@ -8,7 +8,6 @@
 #include "ray/common/task/scheduling_resources.h"
 #include "ray/common/task/task.h"
 #include "ray/common/task/task_common.h"
-#include "ray/rpc/worker/direct_actor_client.h"
 #include "ray/rpc/worker/worker_client.h"
 
 namespace ray {
@@ -107,8 +106,6 @@ class Worker {
   /// Whether the worker is detached. This is applies when the worker is actor.
   /// Detached actor means the actor's creator can exit without killing this actor.
   bool is_detached_actor_;
-  /// The rpc client to send tasks to the direct actor service.
-  std::unique_ptr<rpc::DirectActorClient> direct_rpc_client_;
 };
 
 }  // namespace raylet
